requirements.txt
Flask
twilio
python-dotenv
requests
APScheduler
psycopg2-binary # NOVO: Driver PostgreSQL